backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.89.0",
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "langgraph>=1.1.6",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "spotipy>=2.26.0",
    "uvicorn>=0.44.0",
]
